utils: Move input_* functions to own submodule
