+++ cobbler/items/file.py
@@ -8,7 +8,7 @@ Cobbler module that contains the code for a Cobbler file object.
 
 import uuid
 
-from cobbler import utils
+from cobbler.utils import input_converters
 from cobbler.items import resource
 
 from cobbler.cexceptions import CX
@@ -99,7 +99,7 @@ class File(resource.Resource):
         :param is_dir: This is the path to check if it is a directory.
         :raises TypeError: Raised in case ``is_dir`` is not a boolean.
         """
-        is_dir = utils.input_boolean(is_dir)
+        is_dir = input_converters.input_boolean(is_dir)
         if not isinstance(is_dir, bool):
             raise TypeError("Field is_dir in object file needs to be of type bool!")
         self._is_dir = is_dir
