+++ tests/utils/__init__.py
@@ -209,64 +209,6 @@ def test_file_is_remote(remote_url, expected_result):
     assert expected_result == result
 
 
-@pytest.mark.parametrize(
-    "test_input,expected_result", [("<<inherit>>", "<<inherit>>"), ("delete", [])]
-)
-def test_input_string_or_list(test_input, expected_result):
-    # Arrange
-
-    # Act
-    result = utils.input_string_or_list(test_input)
-
-    # Assert
-    assert expected_result == result
-
-
-@pytest.mark.parametrize(
-    "testinput,expected_result,possible_exception",
-    [
-        ("<<inherit>>", "<<inherit>>", does_not_raise()),
-        ([""], None, pytest.raises(TypeError)),
-        ("a b=10 c=abc", {"a": None, "b": "10", "c": "abc"}, does_not_raise()),
-        ({"ab": 0}, {"ab": 0}, does_not_raise()),
-        (0, None, pytest.raises(TypeError)),
-    ],
-)
-def test_input_string_or_dict(testinput, expected_result, possible_exception):
-    # Arrange
-
-    # Act
-    with possible_exception:
-        result = utils.input_string_or_dict(testinput)
-
-        # Assert
-        assert expected_result == result
-
-
-@pytest.mark.parametrize(
-    "testinput,expected_exception,expected_result",
-    [
-        (True, does_not_raise(), True),
-        (1, does_not_raise(), True),
-        ("oN", does_not_raise(), True),
-        ("yEs", does_not_raise(), True),
-        ("Y", does_not_raise(), True),
-        ("Test", does_not_raise(), False),
-        (-5, does_not_raise(), False),
-        (0.5, pytest.raises(TypeError), False),
-    ],
-)
-def test_input_boolean(testinput, expected_exception, expected_result):
-    # Arrange
-
-    # Act
-    with expected_exception:
-        result = utils.input_boolean(testinput)
-
-        # Assert
-        assert expected_result == result
-
-
 def test_blender(cobbler_api):
     # Arrange
     root_item = Distro(cobbler_api)
