+++ tests/utils/input_converters_test.py
@@ -1,0 +1,63 @@
+import pytest
+
+from tests.conftest import does_not_raise
+
+from cobbler.utils import input_converters
+
+
+@pytest.mark.parametrize(
+    "test_input,expected_result", [("<<inherit>>", "<<inherit>>"), ("delete", [])]
+)
+def test_input_string_or_list(test_input, expected_result):
+    # Arrange
+
+    # Act
+    result = input_converters.input_string_or_list(test_input)
+
+    # Assert
+    assert expected_result == result
+
+
+@pytest.mark.parametrize(
+    "testinput,expected_result,possible_exception",
+    [
+        ("<<inherit>>", "<<inherit>>", does_not_raise()),
+        ([""], None, pytest.raises(TypeError)),
+        ("a b=10 c=abc", {"a": None, "b": "10", "c": "abc"}, does_not_raise()),
+        ({"ab": 0}, {"ab": 0}, does_not_raise()),
+        (0, None, pytest.raises(TypeError)),
+    ],
+)
+def test_input_string_or_dict(testinput, expected_result, possible_exception):
+    # Arrange
+
+    # Act
+    with possible_exception:
+        result = input_converters.input_string_or_dict(testinput)
+
+        # Assert
+        assert expected_result == result
+
+
+@pytest.mark.parametrize(
+    "testinput,expected_exception,expected_result",
+    [
+        (True, does_not_raise(), True),
+        (1, does_not_raise(), True),
+        ("oN", does_not_raise(), True),
+        ("yEs", does_not_raise(), True),
+        ("Y", does_not_raise(), True),
+        ("Test", does_not_raise(), False),
+        (-5, does_not_raise(), False),
+        (0.5, pytest.raises(TypeError), False),
+    ],
+)
+def test_input_boolean(testinput, expected_exception, expected_result):
+    # Arrange
+
+    # Act
+    with expected_exception:
+        result = input_converters.input_boolean(testinput)
+
+        # Assert
+        assert expected_result == result
